+++ src/repo_man/commands/edit.py
@@ -1,16 +1,13 @@
-from pathlib import Path
-
 import click
 
 from repo_man.consts import REPO_TYPES_CFG
+from repo_man.utils import ensure_config_file_exists
 
 
 @click.command
 def edit():
     """Edit the repo-man configuration manually"""
 
-    if not Path(REPO_TYPES_CFG).exists():
-        click.echo(click.style(f"No {REPO_TYPES_CFG} file found.", fg="red"))
-        raise SystemExit(1)
+    ensure_config_file_exists()
 
     click.edit(filename=REPO_TYPES_CFG)