+++ src/repo_man/commands/list_repos.py
@@ -1,10 +1,8 @@
 import configparser
-from pathlib import Path
 
 import click
 
-from repo_man.consts import REPO_TYPES_CFG
-from repo_man.utils import parse_repo_types, pass_config
+from repo_man.utils import parse_repo_types, pass_config, ensure_config_file_exists
 
 
 @click.command(name="list", help="The type of repository to manage")
@@ -13,9 +11,7 @@ from repo_man.utils import parse_repo_types, pass_config
 def list_repos(config: configparser.ConfigParser, repo_types: list[str]):
     """List matching repositories"""
 
-    if not Path(REPO_TYPES_CFG).exists():
-        click.echo(click.style(f"No {REPO_TYPES_CFG} file found.", fg="red"))
-        raise SystemExit(1)
+    ensure_config_file_exists()
 
     valid_repo_types = parse_repo_types(config)
     found_repos = set()