+++ docs/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- `remove` command to remove a repository from one or more types
+
 ## [0.0.7] - 2024-01-21
 
 ### Added